Setup new positioning mechanism
The relationships system was extremely complicated to try to navigate in
code. I realized that if we abandon percentages and instead use a pixel
based system for a 'virtual screen', it makes things much simpler. We can
virtually resize and position each screen and have actual coordinates right
from the config.